requirements.txt
fastapi
uvicorn
pandas
python-multipart
python-dotenv
openai